fix refactor bugs
