+++ src/server/api/handlers/deleters/sprintDeleter.ts
@@ -4,7 +4,7 @@ import { FindOptions, InstanceDestroyOptions, Transaction } from "sequelize";
 
 // utils
 import { buildOptionsWithTransaction } from "../../utils/sequelizeHelper";
-import { mapToSprint } from "../../../dataaccess/mappers/apiToDataAccessMappers";
+import { mapToSprint } from "../../../dataaccess/mappers/dataAccessToApiMappers";
 
 // data access
 import { SprintModel } from "../../../dataaccess/models/Sprint";
