+++ src/server/api/handlers/fetchers/sprintFetcher.ts
@@ -5,7 +5,7 @@ import * as HttpStatus from "http-status-codes";
 import { ApiSprint } from "@atoll/shared";
 
 // utils
-import { mapToSprint } from "../../../dataaccess/mappers/apiToDataAccessMappers";
+import { mapToSprint } from "../../../dataaccess/mappers/dataAccessToApiMappers";
 import { buildOptionsFromParams } from "../../utils/sequelizeHelper";
 import { buildSelfLink } from "../../../utils/linkBuilder";
