+++ src/server/api/handlers/fetchers/userPreferencesFetcher.ts
@@ -2,7 +2,7 @@
 import * as HttpStatus from "http-status-codes";
 
 // utils
-import { mapToUserSettings } from "../../../dataaccess/mappers/apiToDataAccessMappers";
+import { mapToUserSettings } from "../../../dataaccess/mappers/dataAccessToApiMappers";
 
 // data access
 import { UserSettingsModel } from "../../../dataaccess/models/UserSettings";
