+++ src/server/api/handlers/sprintBacklogItems.ts
@@ -7,7 +7,7 @@ import { CreateOptions, Transaction } from "sequelize";
 import { getParamsFromRequest } from "../utils/filterHelper";
 import { buildOptionsFromParams } from "../utils/sequelizeHelper";
 import { respondWithError } from "../utils/responder";
-import { mapSprintBacklogToBacklogItem, mapToSprintBacklogItem } from "../../dataaccess/mappers/apiToDataAccessMappers";
+import { mapSprintBacklogToBacklogItem, mapToSprintBacklogItem } from "../../dataaccess/mappers/dataAccessToApiMappers";
 import { addIdToBody } from "../utils/uuidHelper";
 import { sprintBacklogItemFetcher } from "./fetchers/sprintBacklogItemFetcher";
