+++ src/server/api/handlers/sprints.ts
@@ -13,7 +13,8 @@ import { deleteSprint } from "./deleters/sprintDeleter";
 import { getParamFromRequest, getParamsFromRequest } from "../utils/filterHelper";
 import { addIdToBody } from "../utils/uuidHelper";
 import { respondWithError, respondWithFailedValidation, respondWithItem, respondWithNotFound } from "../utils/responder";
-import { mapFromSprint, ApiToDataAccessMapOptions, mapToSprint } from "../../dataaccess/mappers/apiToDataAccessMappers";
+import { mapFromSprint, ApiToDataAccessMapOptions } from "../../dataaccess/mappers/apiToDataAccessMappers";
+import { mapToSprint } from "../../dataaccess/mappers/dataAccessToApiMappers";
 import { respondedWithMismatchedItemIds } from "../utils/validationResponders";
 import { getInvalidPatchMessage, getPatchedItem } from "../utils/patcher";
